remove unneccessary code
