+++ pkg/idxcdc/types.go
@@ -20,8 +20,6 @@ import (
 	"encoding/hex"
 	"encoding/json"
 	"fmt"
-	"strings"
-	"sync"
 	"time"
 
 	"github.com/google/uuid"
@@ -30,6 +28,7 @@ import (
 	"github.com/matrixorigin/matrixone/pkg/container/types"
 	"github.com/matrixorigin/matrixone/pkg/container/vector"
 	"github.com/matrixorigin/matrixone/pkg/logutil"
+	"github.com/matrixorigin/matrixone/pkg/util/executor"
 	v2 "github.com/matrixorigin/matrixone/pkg/util/metric/v2"
 	"github.com/matrixorigin/matrixone/pkg/vm/engine/readutil"
 	"github.com/tidwall/btree"
@@ -44,18 +43,7 @@ const (
 )
 
 const (
-	CDCDefaultSendSqlTimeout                 = "10m"
-	CDCDefaultRetryTimes                     = -1
-	CDCDefaultRetryDuration                  = 30 * time.Minute
-	CDCDefaultTaskExtra_InitSnapshotSplitTxn = true
-	CDCDefaultTaskExtra_MaxSQLLen            = 4 * 1024 * 1024
-)
-
-const (
-	CDCSinkType_MySQL     = "mysql"
-	CDCSinkType_MO        = "matrixone"
-	CDCSinkType_Console   = "console"
-	CDCSinkType_IndexSync = "indexsync"
+	ConsumerType_IndexSync = 0
 )
 
 const (
@@ -66,129 +54,27 @@ const (
 	CDCPitrGranularity_All     = "*"
 )
 
-const (
-	CDCRequestOptions_Level                = "Level"
-	CDCRequestOptions_Exclude              = "Exclude"
-	CDCRequestOptions_StartTs              = "StartTs"
-	CDCRequestOptions_EndTs                = "EndTs"
-	CDCRequestOptions_SendSqlTimeout       = "SendSqlTimeout"
-	CDCRequestOptions_InitSnapshotSplitTxn = "InitSnapshotSplitTxn"
-	CDCRequestOptions_MaxSqlLength         = "MaxSqlLength"
-	CDCRequestOptions_NoFull               = "NoFull"
-	CDCRequestOptions_ConfigFile           = "ConfigFile"
-)
-
-const (
-	CDCTaskExtraOptions_MaxSqlLength         = CDCRequestOptions_MaxSqlLength
-	CDCTaskExtraOptions_SendSqlTimeout       = CDCRequestOptions_SendSqlTimeout
-	CDCTaskExtraOptions_InitSnapshotSplitTxn = CDCRequestOptions_InitSnapshotSplitTxn
-)
-
-var CDCRequestOptions = []string{
-	CDCRequestOptions_Level,
-	CDCRequestOptions_Exclude,
-	CDCRequestOptions_StartTs,
-	CDCRequestOptions_EndTs,
-	CDCRequestOptions_MaxSqlLength,
-	CDCRequestOptions_SendSqlTimeout,
-	CDCRequestOptions_InitSnapshotSplitTxn,
-	CDCRequestOptions_ConfigFile,
-	CDCRequestOptions_NoFull,
+type DataRetriever interface {
+	Next() (insertBatch *AtomicBatch, deleteBatch *AtomicBatch, noMoreData bool, err error)
+	UpdateWatermark(executor.TxnExecutor, executor.StatementOption) error
+	GetDataType() int8
 }
 
-var CDCTaskExtraOptions = []string{
-	CDCTaskExtraOptions_MaxSqlLength,
-	CDCTaskExtraOptions_SendSqlTimeout,
-	CDCTaskExtraOptions_InitSnapshotSplitTxn,
+type ConsumerInfo struct {
+	ConsumerType int8
+	TableName    string
+	DbName       string
+	IndexName    string
 }
 
-var (
-	EnableConsoleSink = false
-)
+type Consumer interface {
+	Consume(context.Context, DataRetriever) error
+}
 
 type TaskId = uuid.UUID
 
 func NewTaskId() TaskId {
 	return uuid.Must(uuid.NewV7())
-}
-
-// func StringToTaskId(s string) (TaskId, error) {
-// 	return uuid.Parse(s)
-// }
-
-type Reader interface {
-	Run(ctx context.Context, ar *ActiveRoutine)
-	Close()
-}
-
-type TableReader interface {
-	Run(ctx context.Context, ar *ActiveRoutine)
-	Close()
-	Info() *DbTableInfo
-	GetWg() *sync.WaitGroup
-}
-
-// Sinker manages and drains the sql parts
-type Sinker interface {
-	Run(ctx context.Context, ar *ActiveRoutine)
-	Sink(ctx context.Context, data *DecoderOutput)
-	SendBegin()
-	SendCommit()
-	SendRollback()
-	// SendDummy to guarantee the last sql is sent
-	SendDummy()
-	// Error must be called after Sink
-	Error() error
-	ClearError()
-	Reset()
-	Close()
-}
-
-// Sink represents the destination mysql or matrixone
-type Sink interface {
-	Send(ctx context.Context, ar *ActiveRoutine, sqlBuf []byte, needRetry bool) error
-	SendBegin(ctx context.Context) error
-	SendCommit(ctx context.Context) error
-	SendRollback(ctx context.Context) error
-	Close()
-}
-
-type IWatermarkUpdater interface {
-	Run(ctx context.Context, ar *ActiveRoutine)
-	InsertIntoDb(dbTableInfo *DbTableInfo, watermark types.TS) error
-	GetFromMem(dbName, tblName string) types.TS
-	GetFromDb(dbName, tblName string) (watermark types.TS, err error)
-	UpdateMem(dbName, tblName string, watermark types.TS)
-	DeleteFromMem(dbName, tblName string)
-	DeleteFromDb(dbName, tblName string) error
-	SaveErrMsg(dbName, tblName string, errMsg string) error
-}
-
-type ActiveRoutine struct {
-	sync.Mutex
-	Pause  chan struct{}
-	Cancel chan struct{}
-}
-
-func (ar *ActiveRoutine) ClosePause() {
-	ar.Lock()
-	defer ar.Unlock()
-	close(ar.Pause)
-	// can't set to nil, because some goroutines may still be running, when it goes next round loop,
-	// it found the channel is nil, not closed, will hang there forever
-}
-
-func (ar *ActiveRoutine) CloseCancel() {
-	ar.Lock()
-	defer ar.Unlock()
-	close(ar.Cancel)
-}
-
-func NewCdcActiveRoutine() *ActiveRoutine {
-	return &ActiveRoutine{
-		Pause:  make(chan struct{}),
-		Cancel: make(chan struct{}),
-	}
 }
 
 type OutputType int8
@@ -209,15 +95,6 @@ func (t OutputType) String() string {
 	}
 }
 
-type DecoderOutput struct {
-	outputTyp      OutputType
-	noMoreData     bool
-	fromTs, toTs   types.TS
-	checkpointBat  *batch.Batch
-	insertAtmBatch *AtomicBatch
-	deleteAtmBatch *AtomicBatch
-}
-
 type RowType int
 
 const (
@@ -230,54 +107,6 @@ type RowIterator interface {
 	Next() bool
 	Row(ctx context.Context, row []any) error
 	Close()
-}
-
-type DbTableInfo struct {
-	SourceDbId      uint64
-	SourceDbName    string
-	SourceTblId     uint64
-	SourceTblName   string
-	SourceCreateSql string
-
-	SinkDbName  string
-	SinkTblName string
-
-	IdChanged bool
-}
-
-func (info DbTableInfo) String() string {
-	return fmt.Sprintf("%v(%v).%v(%v) -> %v.%v, %v",
-		info.SourceDbName,
-		info.SourceDbId,
-		info.SourceTblName,
-		info.SourceTblId,
-		info.SinkDbName,
-		info.SinkTblName,
-		info.IdChanged,
-	)
-}
-
-func (info DbTableInfo) Clone() *DbTableInfo {
-	return &DbTableInfo{
-		SourceDbId:      info.SourceDbId,
-		SourceDbName:    info.SourceDbName,
-		SourceTblId:     info.SourceTblId,
-		SourceTblName:   info.SourceTblName,
-		SourceCreateSql: info.SourceCreateSql,
-		SinkDbName:      info.SinkDbName,
-		SinkTblName:     info.SinkTblName,
-		IdChanged:       info.IdChanged,
-	}
-}
-
-func (info DbTableInfo) OnlyDiffinTblId(t *DbTableInfo) bool {
-	if info.SourceDbId != t.SourceDbId ||
-		info.SourceDbName != t.SourceDbName ||
-		info.SourceTblName != t.SourceTblName ||
-		info.SourceCreateSql != t.SourceCreateSql {
-		return false
-	}
-	return info.SourceTblId != t.SourceTblId
 }
 
 // AtomicBatch holds batches from [Tail_wip,...,Tail_done] or [Tail_done].
@@ -457,49 +286,6 @@ func (info *UriInfo) GetEncodedPassword() (string, error) {
 
 func (info *UriInfo) String() string {
 	return fmt.Sprintf("%s%s:%s@%s:%d", CDCSourceUriPrefix, info.User, "******", info.Ip, info.Port)
-}
-
-type PatternTable struct {
-	Database string `json:"database"`
-	Table    string `json:"table"`
-}
-
-func (table PatternTable) String() string {
-	return fmt.Sprintf("%s.%s", table.Database, table.Table)
-}
-
-type PatternTuple struct {
-	Source       PatternTable `json:"Source"`
-	Sink         PatternTable `json:"Sink"`
-	OriginString string       `json:"-"`
-	Reserved     string       `json:"reserved"`
-}
-
-func (tuple *PatternTuple) String() string {
-	if tuple == nil {
-		return ""
-	}
-	return fmt.Sprintf("%s,%s", tuple.Source, tuple.Sink)
-}
-
-type PatternTuples struct {
-	Pts      []*PatternTuple `json:"pts"`
-	Reserved string          `json:"reserved"`
-}
-
-func (pts *PatternTuples) Append(pt *PatternTuple) {
-	pts.Pts = append(pts.Pts, pt)
-}
-
-func (pts *PatternTuples) String() string {
-	if pts.Pts == nil {
-		return ""
-	}
-	ss := make([]string, 0)
-	for _, pt := range pts.Pts {
-		ss = append(ss, pt.String())
-	}
-	return strings.Join(ss, ",")
 }
 
 // JsonEncode encodes the object to json
